feat: configurable operation details bottom sheet
