+++ ts/features/idpay/initiative/details/components/TimelineDetailsBottomSheet.tsx
@@ -1,75 +1,25 @@
 import * as pot from "@pagopa/ts-commons/lib/pot";
 import React from "react";
-import { Image, StyleSheet, View } from "react-native";
-import { OperationDTO } from "../../../../../../definitions/idpay/timeline/OperationDTO";
+import { StyleSheet, View } from "react-native";
 import { OperationListDTO } from "../../../../../../definitions/idpay/timeline/OperationListDTO";
-import {
-  OperationTypeEnum as TransactionDetailOperationTypeEnum,
-  TransactionDetailDTO
-} from "../../../../../../definitions/idpay/timeline/TransactionDetailDTO";
-import {
-  OperationTypeEnum as RefundOperationTypeEnum,
-  RefundOperationDTO
-} from "../../../../../../definitions/idpay/timeline/RefundOperationDTO";
+import { OperationTypeEnum as RefundOperationTypeEnum } from "../../../../../../definitions/idpay/timeline/RefundOperationDTO";
+import { OperationTypeEnum as TransactionDetailOperationTypeEnum } from "../../../../../../definitions/idpay/timeline/TransactionDetailDTO";
 import { InitiativeDTO } from "../../../../../../definitions/idpay/wallet/InitiativeDTO";
-import { InfoBox } from "../../../../../components/box/InfoBox";
 import ButtonDefaultOpacity from "../../../../../components/ButtonDefaultOpacity";
-import CopyButtonComponent from "../../../../../components/CopyButtonComponent";
 import { Pictogram } from "../../../../../components/core/pictograms";
-import { HSpacer, VSpacer } from "../../../../../components/core/spacer/Spacer";
-import { Body } from "../../../../../components/core/typography/Body";
 import { H4 } from "../../../../../components/core/typography/H4";
-import { IOStyles } from "../../../../../components/core/variables/IOStyles";
-import ItemSeparatorComponent from "../../../../../components/ItemSeparatorComponent";
 import LoadingSpinnerOverlay from "../../../../../components/LoadingSpinnerOverlay";
 import I18n from "../../../../../i18n";
 import { useIODispatch, useIOSelector } from "../../../../../store/hooks";
 import themeVariables from "../../../../../theme/variables";
-import { format } from "../../../../../utils/dates";
 import {
   IOBottomSheetModal,
   useIOBottomSheetModal
 } from "../../../../../utils/hooks/bottomSheet";
-import { formatNumberAmount } from "../../../../../utils/stringBuilder";
 import { idpayTimelineDetailsSelector } from "../store";
 import { idpayTimelineDetailsGet } from "../store/actions";
-
-// TODO: this is temporary, mapping should be done on the backend
-const operationCircuitTypeMap: Record<string, string> = {
-  "00": "Bancomat",
-  "01": "Visa",
-  "02": "Mastercard",
-  "03": "Amex",
-  "04": "JCB",
-  "05": "UnionPay",
-  "06": "Diners",
-  "07": "PostePay",
-  "08": "BancomatPay",
-  "09": "Satispay",
-  "10": "PrivateCircuit"
-};
-
-type InstrumentDetailsComponentProps = Pick<
-  TransactionDetailDTO,
-  "brandLogo" | "maskedPan"
->;
-
-/**
- * Displays the info of the instrument used for the transaction, with the logo of the brand and the masked pan
- * @param {InstrumentDetailsComponentProps} props
- * @returns
- */
-const InstrumentDetailsComponent = (props: InstrumentDetailsComponentProps) => (
-  <View style={styles.centerRow}>
-    <Image style={styles.brandLogo} source={{ uri: props.brandLogo }} />
-    <HSpacer size={8} />
-    <Body weight="SemiBold">
-      {I18n.t("idpay.initiative.operationDetails.maskedPan", {
-        lastDigits: props.maskedPan
-      })}
-    </Body>
-  </View>
-);
+import { RefundDetailsComponent } from "./RefundDetailsComponent";
+import { TransactionDetailsComponent } from "./TransactionDetailsComponent";
 
 /**
  * Displays a generic error message
@@ -80,127 +30,6 @@ const TimelineDetailsErrorComponent = () => (
     <Pictogram name="error" />
   </View>
 );
-
-type CopyTextProps = {
-  text: string;
-};
-
-/**
- * Utility component to display a text with a CopyButtonComponent
- * @param props
- */
-const CopyTextComponent = (props: CopyTextProps) => (
-  <View style={[IOStyles.flex, IOStyles.row]}>
-    <Body
-      weight="SemiBold"
-      numberOfLines={1}
-      ellipsizeMode="tail"
-      style={IOStyles.flex}
-    >
-      {props.text}
-    </Body>
-    <HSpacer size={8} />
-    <CopyButtonComponent textToCopy={props.text} />
-  </View>
-);
-
-type TransactionDetailsProps = {
-  transaction: TransactionDetailDTO;
-};
-
-/**
- * Displays details for both `REVERSAL` and `TRANSACTION` type operation
- * @param props
- * @returns
- */
-const TransactionDetailsComponent = (props: TransactionDetailsProps) => {
-  const { transaction } = props;
-
-  const isReversal =
-    transaction.operationType === TransactionDetailOperationTypeEnum.REVERSAL;
-
-  return (
-    <View style={styles.container}>
-      {isReversal && (
-        <>
-          <InfoBox>
-            <Body>
-              {I18n.t("idpay.initiative.operationDetails.reversalAdvice")}
-            </Body>
-          </InfoBox>
-          <VSpacer size={16} />
-        </>
-      )}
-      <View style={styles.detailRow}>
-        <Body>{I18n.t("idpay.initiative.operationDetails.instrument")}</Body>
-        <InstrumentDetailsComponent {...transaction} />
-      </View>
-      <View style={styles.detailRow}>
-        <Body>{I18n.t("idpay.initiative.operationDetails.amountLabel")}</Body>
-        <Body weight="SemiBold">
-          {formatNumberAmount(transaction.amount, true)}
-        </Body>
-      </View>
-      <View style={styles.detailRow}>
-        <Body>
-          {I18n.t("idpay.initiative.operationDetails.accruedAmountLabel")}
-        </Body>
-        <Body weight="SemiBold">
-          {formatNumberAmount(transaction.accrued, true)}
-        </Body>
-      </View>
-      <ItemSeparatorComponent noPadded={true} />
-      <VSpacer size={24} />
-      <H4>{I18n.t("idpay.initiative.operationDetails.infoTitle")}</H4>
-      <View style={styles.detailRow}>
-        <Body>{I18n.t("idpay.initiative.operationDetails.date")}</Body>
-        <Body weight="SemiBold">
-          {format(transaction.operationDate, "DD MMM YYYY, HH:mm")}
-        </Body>
-      </View>
-      <View style={styles.detailRow}>
-        <Body>{I18n.t("idpay.initiative.operationDetails.circuit")}</Body>
-        <Body weight="SemiBold">
-          {operationCircuitTypeMap[transaction.circuitType]}
-        </Body>
-      </View>
-      <View style={styles.detailRow}>
-        <Body>{I18n.t("idpay.initiative.operationDetails.acquirerId")}</Body>
-        <HSpacer size={16} />
-        <CopyTextComponent text={transaction.idTrxAcquirer} />
-      </View>
-      <View style={styles.detailRow}>
-        <Body style={{ flex: 1 }}>
-          {I18n.t("idpay.initiative.operationDetails.issuerId")}
-        </Body>
-        <HSpacer size={16} />
-        <CopyTextComponent text={transaction.idTrxIssuer} />
-      </View>
-    </View>
-  );
-};
-
-type TimelineDetailsProps = {
-  operation: OperationDTO;
-};
-
-/**
- *  This component is used to render the details of an operation
- * @param {TransactionDetailsProps} props
- * @returns {React.ReactElement}
- */
-const TimelineDetailsComponent = (props: TimelineDetailsProps) => {
-  const { operation } = props;
-
-  switch (operation.operationType) {
-    case TransactionDetailOperationTypeEnum.TRANSACTION:
-    case TransactionDetailOperationTypeEnum.REVERSAL:
-      return <TransactionDetailsComponent transaction={operation} />;
-    default:
-      // We don't show additional info for other operation types
-      return null;
-  }
-};
 
 /**
  * Component displayed in the bottom sheet to show the details of a timeline operation.
@@ -215,7 +44,19 @@ const TimelineDetailsBottomSheet = () => {
     if (pot.isError(detailsPot)) {
       return <TimelineDetailsErrorComponent />;
     } else if (pot.isSome(detailsPot)) {
-      return <TimelineDetailsComponent operation={detailsPot.value} />;
+      const operation = detailsPot.value;
+
+      switch (operation.operationType) {
+        case TransactionDetailOperationTypeEnum.TRANSACTION:
+        case TransactionDetailOperationTypeEnum.REVERSAL:
+          return <TransactionDetailsComponent transaction={operation} />;
+        case RefundOperationTypeEnum.PAID_REFUND:
+        case RefundOperationTypeEnum.REJECTED_REFUND:
+          return <RefundDetailsComponent refund={operation} />;
+        default:
+          // We don't show additional info for other operation types
+          return null;
+      }
     }
     return null;
   };
@@ -227,11 +68,42 @@ const TimelineDetailsBottomSheet = () => {
   );
 };
 
+type TimelineDetailsBottomSheetConfiguration = {
+  snapPoint: number;
+  title: string;
+};
+
+type OperationWithDetailsTypeEnum =
+  | TransactionDetailOperationTypeEnum
+  | RefundOperationTypeEnum;
+
+const bottomSheetConfigurations: Record<
+  OperationWithDetailsTypeEnum,
+  TimelineDetailsBottomSheetConfiguration
+> = {
+  PAID_REFUND: {
+    snapPoint: 530,
+    title: I18n.t("idpay.initiative.operationDetails.title.refund")
+  },
+  REJECTED_REFUND: {
+    snapPoint: 530,
+    title: I18n.t("idpay.initiative.operationDetails.title.refund")
+  },
+  TRANSACTION: {
+    snapPoint: 530,
+    title: I18n.t("idpay.initiative.operationDetails.title.transaction")
+  },
+  REVERSAL: {
+    snapPoint: 620,
+    title: I18n.t("idpay.initiative.operationDetails.title.transaction")
+  }
+};
+
 export type TimelineDetailsBottomSheetModal = Omit<
   IOBottomSheetModal,
   "present"
 > & {
-  present: (operationId: OperationListDTO["operationId"]) => void;
+  present: (operation: OperationListDTO) => void;
 };
 
 /**
@@ -243,6 +115,9 @@ export const useTimelineDetailsBottomSheet = (
   initiativeId: InitiativeDTO["initiativeId"]
 ): TimelineDetailsBottomSheetModal => {
   const dispatch = useIODispatch();
+
+  const [snapPoint, setSnapPoint] = React.useState(530);
+  const [title, setTitle] = React.useState("");
 
   const bottomSheetFooter = (
     <View style={styles.footer}>
@@ -258,13 +133,30 @@ export const useTimelineDetailsBottomSheet = (
 
   const modal = useIOBottomSheetModal(
     <TimelineDetailsBottomSheet />,
-    I18n.t("idpay.initiative.operationDetails.title"),
-    530,
+    title,
+    snapPoint,
     bottomSheetFooter
   );
 
-  const present = (operationId: OperationListDTO["operationId"]) => {
-    dispatch(idpayTimelineDetailsGet.request({ initiativeId, operationId }));
+  const present = (operation: OperationListDTO) => {
+    const config =
+      bottomSheetConfigurations[
+        operation.operationType as OperationWithDetailsTypeEnum
+      ];
+
+    if (config === undefined) {
+      return;
+    }
+
+    setSnapPoint(config.snapPoint);
+    setTitle(config.title);
+
+    dispatch(
+      idpayTimelineDetailsGet.request({
+        initiativeId,
+        operationId: operation.operationId
+      })
+    );
     modal.present();
   };
 
@@ -274,24 +166,5 @@ export const useTimelineDetailsBottomSheet = (
 const styles = StyleSheet.create({
   footer: {
     padding: themeVariables.contentPadding
-  },
-  container: {
-    flex: 1,
-    flexGrow: 1,
-    paddingVertical: themeVariables.spacerHeight
-  },
-  detailRow: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    paddingTop: themeVariables.spacerSmallHeight,
-    paddingBottom: themeVariables.spacerSmallHeight
-  },
-  centerRow: {
-    flexDirection: "row",
-    alignItems: "center"
-  },
-  brandLogo: {
-    width: 24,
-    height: 16
   }
 });
